+++ search-films/src/pages/FavoritesPage/index.tsx
@@ -11,7 +11,7 @@ function FavoritesPage() {
 			<Content>
 
 				<Title text='Избранное'/>
-				<Flex paddingTop={88}>
+				<Flex position='horizontal' paddingTop={88}>
 					<MovieCard
 						movieName="Black Window"
 						movieId={'1'}